+++ outsystems/exceptions/osptool_error.py
@@ -1,0 +1,2 @@
+class OSPToolDeploymentError(Exception):
+    pass