+++ outsystems/pipeline/deploy_apps_to_target_env_with_airgap.py
@@ -22,18 +22,24 @@ from outsystems.lifetime.lifetime_applications import export_app_oap
 from outsystems.file_helpers.file import load_data
 from outsystems.lifetime.lifetime_base import build_lt_endpoint
 from outsystems.cicd_probe.cicd_base import build_probe_endpoint
-from outsystems.osp_tool.osp_base import deploy_app_oap
+from outsystems.osp_tool.osp_base import call_osptool
 from outsystems.cicd_probe.cicd_dependencies import get_app_dependencies, sort_app_dependencies
-from outsystems.pipeline.deploy_latest_tags_to_target_env import generate_deployment_based_on_manifest, generate_regular_deployment
+from outsystems.pipeline.deploy_latest_tags_to_target_env import generate_deployment_based_on_manifest as generate_deployment_based_on_deploy_manifest, \
+    generate_regular_deployment
+from outsystems.pipeline.deploy_tags_to_target_env_with_manifest import generate_deployment_based_on_manifest as generate_deployment_based_on_trigger_manifest
 # Exceptions
 from outsystems.exceptions.invalid_parameters import InvalidParametersError
 
 
 # ############################################################# SCRIPT ##############################################################
-def generate_oap_list(app_data_list: list):
+def generate_oap_list(app_data_list: list, friendly_package_names: bool):
     app_oap_list = []
+    filename = ""
     for app in app_data_list:
-        filename = "{}{}".format(app["VersionKey"], APPLICATION_OAP_FILE)
+        if friendly_package_names:
+            filename = "{}_v{}{}".format(app["Name"].replace(" ", "_"), app["Version"].replace(".", "_"), APPLICATION_OAP_FILE)
+        else:
+            filename = "{}{}".format(app["VersionKey"], APPLICATION_OAP_FILE)
         app_oap_list.append({"app_name": app["Name"], "app_version": app["Version"], "app_key": app["Key"], "version_key": app["VersionKey"], "filename": filename})
     return app_oap_list
 
@@ -62,10 +68,10 @@ def generate_deployment_order(artifact_dir: str, probe_endpoint: str, app_oap_li
 def deploy_apps_oap(artifact_dir: str, dest_env: str, osp_tool_path: str, credentials: str, app_oap_list: list):
     for app in app_oap_list:
         oap_file_path = os.path.join(artifact_dir, APPLICATION_OAP_FOLDER, app["filename"])
-        deploy_app_oap(osp_tool_path, oap_file_path, dest_env, credentials)
+        call_osptool(osp_tool_path, oap_file_path, dest_env, credentials)
 
 
-def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: str, lt_api_version: int, lt_token: str, source_env: str, dest_env: str, apps: list, dep_manifest: list, dep_note: str, osp_tool_path: str, credentials: str, cicd_http_proto: str, cicd_url: str, cicd_api_endpoint: str, cicd_version: str):
+def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: str, lt_api_version: int, lt_token: str, source_env: str, dest_env: str, apps: list, dep_manifest: list, trigger_manifest: dict, include_test_apps: bool, dep_note: str, osp_tool_path: str, credentials: str, cicd_http_proto: str, cicd_url: str, cicd_api_endpoint: str, cicd_version: str, friendly_package_names: bool):
 
     app_data_list = []  # will contain the applications to deploy details from LT
 
@@ -84,12 +90,14 @@ def main(artifact_dir: str, lt_http_proto: str, lt_url: str, lt_api_endpoint: st
     # Or else you might not be deploying the same app versions that were deployed in
     # previous pipeline steps
     if dep_manifest:
-        app_data_list = generate_deployment_based_on_manifest(artifact_dir, lt_endpoint, lt_token, src_env_key, source_env, apps, dep_manifest)
+        app_data_list = generate_deployment_based_on_deploy_manifest(artifact_dir, lt_endpoint, lt_token, src_env_key, source_env, apps, dep_manifest)
+    elif trigger_manifest:
+        app_data_list = generate_deployment_based_on_trigger_manifest(artifact_dir, lt_endpoint, lt_token, src_env_key, source_env, trigger_manifest, include_test_apps, False)
     else:
         app_data_list = generate_regular_deployment(artifact_dir, lt_endpoint, lt_token, src_env_key, apps)
 
     # Export binary files
-    app_oap_list = generate_oap_list(app_data_list)
+    app_oap_list = generate_oap_list(app_data_list, friendly_package_names)
     export_apps_oap(artifact_dir, lt_endpoint, lt_token, src_env_key, app_oap_list)
 
     # Generate deployment order
@@ -117,17 +125,19 @@ if __name__ == "__main__":
     parser.add_argument("-v", "--lt_api_version", type=int, default=LIFETIME_API_VERSION,
                         help="LifeTime API version number. If version <= 10, use 1, if version >= 11, use 2. Default: 2")
     parser.add_argument("-e", "--lt_endpoint", type=str, default=LIFETIME_API_ENDPOINT,
-                        help="(optional) Used to set the API endpoint for LifeTime, without the version. Default: \"lifetimeapi/rest\"")
+                        help="(Optional) Used to set the API endpoint for LifeTime, without the version. Default: \"lifetimeapi/rest\"")
     parser.add_argument("-s", "--source_env", type=str, required=True,
                         help="Name, as displayed in LifeTime, of the source environment where the apps are.")
     parser.add_argument("-d", "--destination_env", type=str, required=True,
                         help="Name, as displayed in LifeTime, of the destination environment where you want to deploy the apps. (if in Airgap mode should be the hostname of the destination environment where you want to deploy the apps)")
     parser.add_argument("-m", "--deploy_msg", type=str, default=DEPLOYMENT_MESSAGE,
                         help="Message you want to show on the deployment plans in LifeTime. Default: \"Automated deploy using OS Pipelines\".")
-    parser.add_argument("-l", "--app_list", type=str, required=True,
+    parser.add_argument("-l", "--app_list", type=str,
                         help="Comma separated list of apps you want to deploy. Example: \"App1,App2 With Spaces,App3_With_Underscores\"")
     parser.add_argument("-f", "--manifest_file", type=str,
-                        help="(optional) Manifest file path, used to promote the same application versions throughout the pipeline execution.")
+                        help="Manifest file path (either deployment or trigger), used to promote the same application versions throughout the pipeline execution.")
+    parser.add_argument("-i", "--include_test_apps", action='store_true',
+                        help="(Optional) Flag that indicates if applications marked as \"Test Application\" in the trigger manifest are included in the Air Gap deployment.")
     parser.add_argument("-o", "--osp_tool_path", type=str, required=True,
                         help="OSP Tool file path")
     parser.add_argument("-user", "--airgap_user", type=str, required=True,
@@ -137,9 +147,11 @@ if __name__ == "__main__":
     parser.add_argument("-pu", "--cicd_probe_url", type=str, required=True,
                         help="URL of the environment where the CI/CD Probe is installed (without the API endpoint).")
     parser.add_argument("-pv", "--cicd_probe_version", type=str, default=PROBE_API_VERSION,
-                        help="(optional) CI/CD Probe API version number.")
+                        help="(Optional) CI/CD Probe API version number.")
     parser.add_argument("-pe", "--cicd_probe_endpoint", type=str, default=PROBE_API_ENDPOINT,
-                        help="(optional) Used to set the API endpoint for CI/CD Probe, without the version.")
+                        help="(Optional) Used to set the API endpoint for CI/CD Probe, without the version.")
+    parser.add_argument("-n", "--friendly_package_names", action='store_true',
+                        help="Flag that indicates if downloaded application packages should have a user-friendly name. Example: \"AppName_v1_2_1\"")
 
     args = parser.parse_args()
 
@@ -166,14 +178,27 @@ if __name__ == "__main__":
     source_env = args.source_env
     # Parse Destination Environment
     dest_env = args.destination_env
-    # Parse App list
-    _apps = args.app_list
-    apps = _apps.split(',')
     # Parse Manifest file if it exists
+    # Based on the file content it can be a deployment manifest (list-based) or trigger manifest (dict-based)
     if args.manifest_file:
         manifest_file = load_data("", args.manifest_file)
     else:
         manifest_file = None
+    dep_manifest = manifest_file if type(manifest_file) is list else None
+    trigger_manifest = manifest_file if type(manifest_file) is dict else None
+    # Check if either an app list or a manifest file is being provided
+    if not args.app_list and not args.manifest_file:
+        raise InvalidParametersError("either --app_list or --manifest_file must be provided as arguments")
+
+    if dep_manifest and not args.app_list:
+        raise InvalidParametersError("--app_list parameter is required for Deployment Manifest operation")
+    # Parse App list
+    apps = None
+    if args.app_list:
+        _apps = args.app_list
+        apps = _apps.split(',')
+    # Parse Include Test Apps flag
+    include_test_apps = args.include_test_apps
     # Parse Deployment Message
     dep_note = args.deploy_msg
     # Parse OSP Tool path
@@ -195,6 +220,8 @@ if __name__ == "__main__":
         cicd_url = cicd_url[:-1]
     # Parse CICD Probe API Version
     cicd_version = args.cicd_probe_version
+    # Parse Friendly Package Names flag
+    friendly_package_names = args.friendly_package_names
 
     # Calls the main script
-    main(artifact_dir, lt_http_proto, lt_url, lt_api_endpoint, lt_version, lt_token, source_env, dest_env, apps, manifest_file, dep_note, osp_tool_path, credentials, cicd_http_proto, cicd_url, cicd_api_endpoint, cicd_version)
+    main(artifact_dir, lt_http_proto, lt_url, lt_api_endpoint, lt_version, lt_token, source_env, dest_env, apps, dep_manifest, trigger_manifest, include_test_apps, dep_note, osp_tool_path, credentials, cicd_http_proto, cicd_url, cicd_api_endpoint, cicd_version, friendly_package_names)